backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
opencv-python==4.9.0.80
mediapipe==0.10.21
python-dotenv==1.0.0
aiofiles==23.2.1